[ADD] add publisher API

+++ src/main/java/com/subrutin/catalog/dto/ResultPageResponseDTO.java
@@ -19,7 +19,7 @@ public class ResultPageResponseDTO<T> implements Serializable {
 
 	private List<T> result;
 	
-	private Long pages;
+	private Integer pages;
 	
 	private Long elements;
